First implementation of the new post sharing code.
